requirements.txt
python-bidi
bs4
